Added Index.html for add-to-cart component
The component has been updated to include internal html/css and JavaScript the component have been updated to include interactive elements for adding and removing items to the cart it will also alert users when they do not have any books in the cart when clicking the cart sign in the upper right part of the site. The site is currently a demo but there are chances it may be revisited to become a full functional repo for buying books and working links but it will be later when structure and functionality is planned.